# Change rollout settings for Fluentd to speed up rollout.