Fluentd takes ~1m to boot. With the old settings, a rollout of fluentd on a large cluster can take very long, while there is no need for fluentd to be available.

```
tigera-fluentd                 fluentd-node-nlf7f                              1/1     Terminating       8 (24m ago)     15d
tigera-fluentd                 fluentd-node-r22cf                              1/1     Terminating       8 (24m ago)     15d
tigera-fluentd                 fluentd-node-lx65p                              1/1     Terminating       8 (24m ago)     15d
tigera-fluentd                 fluentd-node-x6b4f                              1/1     Terminating       8 (24m ago)     15d
tigera-fluentd                 fluentd-node-nlf7f                              1/1     Terminating       8 (24m ago)     15d
tigera-fluentd                 fluentd-node-r22cf                              1/1     Terminating       8 (24m ago)     15d
tigera-fluentd                 fluentd-node-lx65p                              1/1     Terminating       8 (24m ago)     15d
tigera-fluentd                 fluentd-node-x6b4f                              1/1     Terminating       8 (24m ago)     15d
tigera-fluentd                 fluentd-node-9j85c                              0/1     Pending           0               0s
tigera-fluentd                 fluentd-node-rgcpr                              0/1     Pending           0               0s
tigera-fluentd                 fluentd-node-x5hwr                              0/1     Pending           0               0s
tigera-fluentd                 fluentd-node-9j85c                              0/1     Pending           0               0s
tigera-fluentd                 fluentd-node-rgcpr                              0/1     Pending           0               0s
tigera-fluentd                 fluentd-node-tpms6                              0/1     Pending           0               0s
tigera-fluentd                 fluentd-node-x5hwr                              0/1     Pending           0               0s
tigera-fluentd                 fluentd-node-tpms6                              0/1     Pending           0               0s
tigera-fluentd                 fluentd-node-9j85c                              0/1     ContainerCreating   0               0s
tigera-fluentd                 fluentd-node-rgcpr                              0/1     ContainerCreating   0               0s
tigera-fluentd                 fluentd-node-x5hwr                              0/1     ContainerCreating   0               1s
tigera-fluentd                 fluentd-node-tpms6                              0/1     ContainerCreating   0               1s
tigera-fluentd                 fluentd-node-9j85c                              0/1     ContainerCreating   0               1s
tigera-fluentd                 fluentd-node-rgcpr                              0/1     ContainerCreating   0               1s
tigera-fluentd                 fluentd-node-x5hwr                              0/1     ContainerCreating   0               1s
...
tigera-fluentd                 fluentd-node-s7cr8                              0/1     Running     0               45s
tigera-fluentd                 fluentd-node-tzwwh                              0/1     Running     0               45s
tigera-fluentd                 fluentd-node-v5qcs                              0/1     Running     0               45s
tigera-fluentd                 fluentd-node-wqjj9                              0/1     Running     0               45s
...
```